{acquisition,behavior}.py: move behavior into acquisition & zap behavior
... these classes were more acquisition related...

also, previous 'DataDescription' file ingest provenance class has been removed.
