+++ acquisition.py
@@ -2,25 +2,30 @@ import datajoint as dj
 
 schema = dj.schema(dj.config['names.%s' % __name__], locals())
 
+'''
+To simplify from Alyx schema, dropping file repositories as core
+parts of tables and moving to per-table blobs. A facility for
+tracking data provenance could still made avaialble via a table
+linked into individual model records as a non-primary attribute.
 
-# To simplify from Alyx schema,
-# dropping file repositories as core parts of tables and moving
-# to per-table blobs. A facility for tracking data provenance is still
-# made avaialble via 'DataDescription' which can be linked into individual
-# model records as a non-primary attribute.
-#
-# This is probably not ideal w/r/t Alyx -
-# one idea would be to merge with external storage when it is available:
-#
-# see also: https://github.com/datajoint/datajoint-python/issues/204
-#
-# with example field def being:
-#
-#   aligned_movie :  external  # motion-aligned movie
-#
-# And have a separate 'phase 1' data schema,
-# which is then coupled with phase 2+ for higher level data products.
+This is probably not ideal w/r/t Alyx -
 
+one idea would be to merge use external storage with example
+field def being:
+
+  aligned_movie :  external  # motion-aligned movie
+
+and also have a separate 'phase 1' data schema, which is then coupled
+with phase 2+ for higher level data products.
+
+Also note: TimeScale items missing wrt TimeSeries items, which were:
+
+- PupilTracking.xyd
+- PupilTracking.movie
+- HeadTracking.xy_theta
+- HeadTracking.movie
+
+TimeScale not yet defined
 
 # SKIPPED:
 # <class 'data.models.DataRepositoryType'>
@@ -33,17 +38,87 @@ schema = dj.schema(dj.config['names.%s' % __name__], locals())
 # <class 'data.models.TimeSeries'>
 # <class 'data.models.EventSeries'>
 # <class 'data.models.IntervalSeries'>
+'''
 
 
 @schema
-class DataDescription(dj.Manual):
-    # FIXME: SKIPPING for now.
-    # <class 'data.models.BaseExperimentalData> # Abstract
+class Session(dj.Manual):
+    # <class 'actions.models.Session'>
+    # XXX: session_type table?
+    definition = """
+    -> subject.Subject
+    session_number:             integer		# number
+    session_start_time:         datetime	# start time
+    ---
+    session_end_time:           datetime	# end time
+    session_type:		varchar(255)	# type
+    -> equipment.LabLocation
+    -> reference.User
+    """
+
+
+@schema
+class PupilTracking(dj.Manual):
+    # <class 'behavior.models.PupilTracking'>
     definition = """
     -> Session
-    -> User
-    datadescription_id:		char(32)		# data id
-    created:			datetime		# created
-    generating_software:	varchar(255)		# generating software
-    provenance_directory:	varchar(255)		# provenance directory
+    pupil_tracking_start_time:  datetime        # start time
+    eye:                        enum("L", "R")  # eye
+    ---
+    pupil_tracking_movie:       longblob        # pupil tracking movie (raw)
+    pupil_trace:                longblob        # x y d
+    """
+
+
+@schema
+class HeadTracking(dj.Manual):
+    # <class 'behavior.models.HeadTracking'>
+    definition = """
+    -> Session
+    head_tracking_start_time:   datetime        # start time
+    ---
+    head_tracking_movie:        longblob        # head tracking movie (raw)
+    head_position:              longblob        # x y theta
+    """
+
+
+@schema
+class OptogeneticStimulus(dj.Manual):
+    # <class 'behavior.models.OptogeneticStimulus'>
+    definition = """
+    -> Session
+    -> reference.BrainLocation
+    ---
+    -> equipment.LightSource
+    light_delivery:		varchar(255)	# light delivery
+    wavelength:    		float		# wavelength
+    power_calculation_method:	varchar(255)	# power calculation method
+    description:		varchar(255)	# description
+    """
+
+    class Pulse(dj.Part):
+        definition = """
+        -> OptogeneticStimulus
+        pulse_start_time:       float           # pulse start time
+        ---
+        pulse_stop_time:        float           # pulse stop time
+        pulse_power:            smallint        # pulse power
+        pulse_x:                float           # pulse x position
+        pulse_y:                float           # pulse y position
+        pulse_z:                float           # pulse z position
+        """
+
+
+@schema
+class Pharmacology(dj.Manual):
+    # <class 'behavior.models.Pharmacology'>
+    definition = """
+    -> Session
+    drug:			varchar(255)	# drug
+    administration_start_time:  float		# start time
+    ---
+    administration_end_time:    float		# end time
+    administration_route:	varchar(255)	# administration route
+    concentration:		float           # concentration
+    volume:			float   	# volume
     """
